requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.25.7
anyio==4.3.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
dataclasses-json==0.6.5
defusedxml==0.7.1
distro==1.9.0
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.3.1
google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.7
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.17
langchain-anthropic==0.1.11
langchain-community==0.0.36
langchain-core==0.1.50
langchain-experimental==0.0.57
langchain-google-genai==1.0.3
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
langsmith==0.1.54
marshmallow==3.21.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.25.1
orjson==3.10.3
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.1.2
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
rsa==4.9
setuptools==68.2.2
sniffio==1.3.1
SQLAlchemy==2.0.29
tenacity==8.2.3
tiktoken==0.6.0
tokenizers==0.19.1
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
wheel==0.43.0
yarl==1.9.4
